src/templates/requirements.txt
torch>=1.8.0
pytorch-ignite>=0.4.4
